+++ qom/ui/figure.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
  
-"""Wrapper modules to display matplotlib plots."""
+"""UI module to display matplotlib plots."""
 
 __name__    = 'qom.ui.figure'
 __authors__ = ['Sampreet Kalita']
 __created__ = '2020-06-16'
-__updated__ = '2020-09-23'
+__updated__ = '2020-09-27'
 
 # dependencies
 from matplotlib.lines import Line2D
@@ -15,8 +15,13 @@ import matplotlib.pyplot as plt
 import numpy as np
 import seaborn as sns
 
+# # dev dependencies
+# from qom.utils import misc
+
 # module logger
 logger = logging.getLogger(__name__)
+
+# TODO: handle contour, contourf plots.
 
 class Plotter():
     """Class containing various 2D plot scenarios.
@@ -44,10 +49,13 @@ class Plotter():
                 pcolormesh: Color plot.
                 scatter: Scatter plot.
                 scatters: Multi-scatter plot.
+        
+        fonts : dict
+            Dictionaries of font settings for text and mathematical symbols.
+            
+        labels : dict
+            Dictionaries of axes labels.
     """
-
-    # TODO: handle contour, contourf plots.
-    # TODO: validate multi-scatter plots.
 
     def __init__(self, plot_params, X, Y=None, Z=None):
         """Class constructor for Plotter2D.
@@ -67,30 +75,34 @@ class Plotter():
                 Z-axis data.
         """
         
-        # initialize
-        plt.show()
-        self.axes = plt.gca()
-        self.plot_type = plot_params['type']
+        # initialize plot
+        self.__init_plot(plot_params)
             
         # single-line plot
         if self.__plot_type == 'line':
-            self.__init_1D(X.values, 1, [plot_params['color']], linestyles=[plot_params['linestyle']], legends=[plot_params['legend']])
+            colors = [plot_params['color'] if 'color' in plot_params else 'r']
+            linestyles = [plot_params['linestyle'] if 'linestyle' in plot_params else '-']
+            legends = [plot_params['legend'] if 'legend' in plot_params else '']
+            self.__init_1D(X.values, 1, colors, linestyles=linestyles, legends=legends)
         # multi-line plot
         elif self.__plot_type == 'lines':
             self.__init_1D(X.values, len(Z.legends), Z.colors, linestyles=Z.linestyles, legends=Z.legends)
         # scatter plot
         elif self.__plot_type == 'scatter':
-            self.__init_1D(X.values, 1, [plot_params['color']], sizes=[plot_params['size']], legends=[plot_params['legend']])
+            colors = [plot_params['color'] if 'color' in plot_params else 'r']
+            sizes = [plot_params['size'] if 'size' in plot_params else 2]
+            legends = [plot_params['legend'] if 'legend' in plot_params else '']
+            self.__init_1D(X.values, 1, colors, sizes=sizes, legends=legends)
         # multi-scatter plot
         elif self.__plot_type == 'scatters':
             self.__init_1D(X.values, len(Z.legends), Z.colors, sizes=Z.sizes, legends=Z.legends)
 
+        # contourf plot
+        elif self.plot_type == 'contourf':
+            self.__init_2D(X.values, Y.values, plot_params)
         # pcolormesh plot
         elif self.plot_type == 'pcolormesh':
-            self.__init_2D(X.values, Y.values)
-
-        # setup plot
-        self.__init_params(plot_params)
+            self.__init_2D(X.values, Y.values, plot_params)
 
     @property
     def axes(self):
@@ -226,6 +238,120 @@ class Plotter():
 
         self.__plot_type = plot_type
 
+    @property
+    def fonts(self):
+        """Property fonts.
+
+        Returns
+        -------
+            fonts : dict
+                Dictionaries of font settings for text and mathematical symbols.
+        """
+
+        return self.__fonts
+    
+    @fonts.setter
+    def fonts(self, fonts):
+        """Setter for fonts.
+
+        Returns
+        -------
+            fonts : dict
+                Dictionaries of font settings for text and mathematical symbols.
+        """
+
+        self.__fonts = fonts
+
+    @property
+    def labels(self):
+        """Property labels.
+
+        Returns
+        -------
+            labels : dict
+                Dictionaries of axes labels.
+        """
+
+        return self.__labels
+    
+    @labels.setter
+    def labels(self, labels):
+        """Setter for fonts.
+
+        Returns
+        -------
+            labels : dict
+                Dictionaries of axes labels.
+        """
+
+        self.__labels = labels
+
+    def __init_plot(self, plot_params):
+        """Function to initialize plot with given parameters.
+        
+        Parameters
+        ----------
+            plot_params : dict
+                Parameters for the plot.
+        """
+
+        # get axes
+        plt.show()
+        self.axes = plt.gca()
+        self.plot_type = plot_params['type']
+            
+        # font for texts
+        if 'font_text' in plot_params:
+            _font_text = plot_params['font_text']
+        else:
+            _font_text = {
+                'family': 'Times New Roman',
+                'style': 'normal',
+                'variant': 'normal',
+                'weight': 500,
+                'size': 16.0
+            }
+
+        # font for mathematical symbols
+        if 'font_math' in plot_params:
+            _font_math = plot_params['font_math']
+        else:
+            _font_math = 'cm'
+        plt.rcParams['mathtext.fontset'] = _font_math
+
+        # set fonts property
+        self.fonts = {
+            'text': _font_text,
+            'math': _font_math
+        }
+
+        # title
+        if 'title' in plot_params:
+            plt.title(plot_params['title'], fontdict=_font_text)
+
+        # ticks
+        plt.xticks(fontfamily=_font_text['family'], fontstyle=_font_text['style'], fontvariant=_font_text['variant'], fontweight=_font_text['weight'], fontsize=12.0)
+        plt.yticks(fontfamily=_font_text['family'], fontstyle=_font_text['style'], fontvariant=_font_text['variant'], fontweight=_font_text['weight'], fontsize=12.0)
+        plt.ticklabel_format(axis='both', style='plain')
+
+        # labels
+        x_label = plot_params['x_label'] if 'x_label' in plot_params else ''
+        plt.xlabel(r'' + x_label, fontdict=_font_text)
+        y_label = plot_params['y_label'] if 'y_label' in plot_params else ''
+        plt.ylabel(r'' + y_label, fontdict=_font_text)
+        z_label = plot_params['z_label'] if 'z_label' in plot_params else ''
+        self.labels = {
+            'x_label': x_label,
+            'y_label': y_label,
+            'z_label': z_label
+        }
+
+        # limits
+        if 'x_lim' in plot_params:
+            plt.xlim(plot_params['x_lim'][0], plot_params['x_lim'][1])
+        if 'y_lim' in plot_params:
+            plt.ylim(plot_params['y_lim'][0], plot_params['y_lim'][1])
+
     def __init_1D(self, xs, num, colors, linestyles=None, sizes=None, legends=None):
         """Function to initialize 1D plots.
         
@@ -270,9 +396,16 @@ class Plotter():
 
         # legends
         if legends and legends[0] != '':
-            plt.legend(legends, loc='best')
+            l = plt.legend(legends, loc='best')                
+            plt.setp(l.texts, 
+                fontfamily=self.__fonts['text']['family'], 
+                fontstyle=self.__fonts['text']['style'], 
+                fontvariant=self.__fonts['text']['variant'], 
+                fontweight=self.__fonts['text']['weight'], 
+                fontsize=12.0
+            )
 
-    def __init_2D(self, xs, ys, color_grad='br_light', shading='gouraud', cbar=True):
+    def __init_2D(self, xs, ys, plot_params, color_grad='br_light', shading='gouraud', cbar=True):
         """Function to initialize 2D plots.
         
         Parameters
@@ -282,6 +415,9 @@ class Plotter():
                 
             ys : list
                 Y-axis values.
+                
+            plot_params : dict
+                Parameters for the plot.
 
             color_grad : str, optional
                 Colors for the plot.
@@ -314,42 +450,29 @@ class Plotter():
         if self.__plot_type == 'pcolormesh':
             self.plot = self.__axes.pcolormesh(_X, _Y, _nans, shading=shading, cmap=cmap)
 
+        # pcolormesh plot
+        if self.__plot_type == 'contourf':
+            self.plot = self.__axes.contourf(_X, _Y, _nans, cmap=cmap)
+
         # color bar
         if cbar:
             self.cbar = plt.colorbar(self.__plot)
-
-    def __init_params(self, plot_params):
-        """Function to initialize plot with given parameters.
-        
-        Parameters
-        ----------
-            plot_params : dict
-                Parameters for the plot.
-        """
-
-        # font sizes
-        plt.rcParams.update({'font.size': 12})
-        plt.xticks(fontsize=12)
-        plt.yticks(fontsize=12)
-
-        # title
-        if 'title' in plot_params:
-            plt.title(plot_params['title'])
-
-        # labels
-        if 'x_label' in plot_params:
-            plt.xlabel(r'' + plot_params['x_label'], fontsize=16)
-        if 'y_label' in plot_params:
-            plt.ylabel(r'' + plot_params['y_label'], fontsize=16)
-
-        # limits
-        if 'x_lim' in plot_params:
-            plt.xlim(plot_params['x_lim'][0], plot_params['x_lim'][1])
-        if 'y_lim' in plot_params:
-            plt.ylim(plot_params['y_lim'][0], plot_params['y_lim'][1])
-
-        # ticks
-        plt.ticklabel_format(axis='both', style='plain')
+            self.__cbar.set_label(
+                label=self.__labels['z_label'], 
+                fontfamily=self.__fonts['text']['family'], 
+                fontstyle=self.__fonts['text']['style'], 
+                fontvariant=self.__fonts['text']['variant'], 
+                fontweight=self.__fonts['text']['weight'], 
+                fontsize=self.__fonts['text']['size']
+            )
+            plt.setp(
+                self.__cbar.ax.get_yticklabels(), 
+                fontfamily=self.__fonts['text']['family'], 
+                fontstyle=self.__fonts['text']['style'], 
+                fontvariant=self.__fonts['text']['variant'], 
+                fontweight=self.__fonts['text']['weight'], 
+                fontsize=12.0
+            )
 
     def update(self, X=None, Y=None, Z=None, head=True, hold=False):
         """Function to update plot.
@@ -386,8 +509,10 @@ class Plotter():
             self.__update_1D(X.values, Y.values, head=head)
         
         # 2D plot
-        elif self.__plot_type == 'pcolormesh':
-            self.__update_2D(Z)
+        if self.__plot_type == 'contourf':
+            self.__update_2D(Z.values)
+        if self.__plot_type == 'pcolormesh':
+            self.__update_2D(Z.values)
 
         # draw data
         plt.draw()
@@ -435,35 +560,52 @@ class Plotter():
                 self.__plot[j].set_offsets(XY)
                 
         # handle nan values for limits
-        minis = []
-        maxis = []
+        _minis = []
+        _maxis = []
         for j in range(len(ys)):
             # calculate minimum and maximum values
             if len(ys[j]) != 0:
                 # handle NaN values
                 _no_nan = [y if y == y else 0 for y in ys[j]]
-                minis.append(min(_no_nan))
-                maxis.append(max(_no_nan))
-        self.__axes.set_ylim(min(minis), max(maxis))
 
-    def __update_2D(self, Z):
+                # update limits
+                # _mini, _maxi, _prec = misc.get_limits(min(_no_nan), max(_no_nan))
+                _minis.append(min(_no_nan))
+                _maxis.append(max(_no_nan))
+
+        # set limits
+        self.__axes.set_ylim(min(_minis), max(_maxis))
+
+    def __update_2D(self, zs):
         """Function to udpate 2D plots.
         
         Parameters
         ----------
-            Z : :class:`qom.utils.axis.StaticAxis` or :class:`qom.utils.axis.DynamicAxis`
-                Z-axis data.
+            zs : list
+                Z-axis values.
         """
         
         # update pcolormesh plot
-        rave = np.ravel(Z.values)
-        self.__plot.set_array(rave)
+        rave = np.ravel(zs)
+
+        # contourf plot
+        if self.__plot_type == 'contourf':
+            self.__axes.clear()
+            self.__axes.contourf(zs)
+        # pcolormesh plot
+        if self.__plot_type == 'pcolormesh':
+            self.__plot.set_array(rave)
 
         # handle NaN values
         _no_nan = [z if z == z else 0 for z in rave]
-        self.__plot.set_clim(vmin=min(_no_nan), vmax=max(_no_nan))
+
+        # set limits
+        # _mini, _maxi, _prec = misc.get_limits(min(_no_nan), max(_no_nan), res=1)
+        _mini, _maxi = min(_no_nan), max(_no_nan)
+        self.__plot.set_clim(vmin=_mini, vmax=_maxi)
 
         # color bar
-        self.__cbar.set_ticks(np.linspace(min(_no_nan), max(_no_nan), 11))
+        _ticks = np.linspace(_mini, _maxi, 11)
+        self.__cbar.set_ticks(_ticks)
         self.__cbar.ax.set_autoscale_on(True)
         self.__cbar.draw_all()