+++ qom_legacy/ui/figure.py
@@ -3,10 +3,10 @@
  
 """Wrapper modules to display matplotlib plots."""
 
-__name__    = 'qom.legacy.ui.figure'
+__name__    = 'qom_legacy.ui.figure'
 __authors__ = ['Sampreet Kalita']
 __created__ = '2020-06-16'
-__updated__ = '2020-09-23'
+__updated__ = '2020-09-27'
 
 # dependencies
 from matplotlib.lines import Line2D