+++ qom_legacy/wrappers/dynamics.py
@@ -3,10 +3,10 @@
 
 """Wrapper modules for dynamics."""
 
-__name__    = 'qom.legacy.wrappers.dynamics'
+__name__    = 'qom_legacy.wrappers.dynamics'
 __authors__ = ['Sampreet Kalita']
 __created__ = '2020-05-01'
-__updated__ = '2020-09-23'
+__updated__ = '2020-09-27'
 
 # dependencies
 import logging
@@ -15,8 +15,8 @@ import os
 import scipy.integrate as si
 
 # dev dependencies
-from qom.measures import corr, diff
-from qom.legacy.ui import figure
+from qom.measures import correlations, differences
+from qom_legacy.ui import figure
 
 # module logger
 logger = logging.getLogger(__name__)
@@ -155,9 +155,9 @@ def dynamics_measure(model, dyna_params, meas_params, plot=False, plot_params=No
     if not found_saved_d:
         # calculate measures
         if meas_params['type'] == 'corr':
-            D = corr.calculate(V, meas_params)
+            D = correlations.calculate(V, meas_params)
         elif meas_params['type'] == 'diff':
-            D = diff.calculate(V, meas_params)
+            D = differences.calculate(V, meas_params)
             
         # display completion
         logger.debug('----------------Measure Dynamics Obtained---------------\n')